FIX: add category to selected product

+++ src/vendor_app/app/src/main/kotlin/cz/quanti/android/vendor_app/repository/product/dto/api/ProductApiEntity.kt
@@ -6,6 +6,6 @@ data class ProductApiEntity(
     var image: String = "",
     var unit: String? = "",
     var categoryId: Long = 0,
-    var unitPrice: Long? = 0,
+    var unitPrice: Double? = 0.0,
     var currency: String? = ""
 )
